chore: prevent vitest mixed exports warning
